analytical_service/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
fonttools==4.28.2
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.2
PyJWT==2.3.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
urllib3==1.26.7
